move consistent hash policy to its own module
